commerce/requirements.txt
amqp==2.6.1
APScheduler==3.7.0
asgiref==3.3.1
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-taggit==1.3.0
django-timezone-field==4.1.2
dnspython==1.16.0
eventlet==0.30.2
flower==0.9.7
greenlet==1.0.0
humanize==3.4.1
idna==2.10
kombu==4.6.11
oauthlib==3.1.0
Pillow==8.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytils==0.3
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tornado==6.1
twilio==6.56.0
tzlocal==2.1
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
